lab8_skeleton/lab8/test_lab8.py
import pytest
from lab8 import convolution2d
import numpy as np
from numpy.testing import assert_array_equal
import pandas as pd

def test_convolution2d():
    # x = np.array([[1,0,0,0], [1,1,1,1], [0,1,0,0], [1,1,0,1]])
    # kernel = np.array([[1,0,0], [0,1,1], [1,1,1]])
    # assert convolution2d(x, kernel) == [[4., 3.], [4., 3.]]

    # x = np.array([[1,1,0,1,1], [0,1,0,0,1], [1,1,0,1,0], [0,0,1,1,1], [1,1,1,1,1]])
    # kernel = np.array([[1,0,1], [1,1,1], [0,1,0]])
    # assert assert_array_equal(convolution2d(x,kernel,stride=2),
    #                          print(np.array([[3., 3.], [3., 4.]])))
    assert True